Group by search

+++ backend/src/main/java/com/devsuperior/dsvendas/dto/SaleDTO.java
@@ -5,19 +5,19 @@ import java.time.LocalDate;
 import com.devsuperior.dsvendas.entities.Sale;
 
 public class SaleDTO {
-	
+
 	private Long id;
 	private Integer visited;
 	private Integer deals;
-	private Integer amount;
+	private Double amount;
 	private LocalDate date;
 	
 	private SellerDTO seller;
 	
-	public SaleDTO() {		
+	public SaleDTO() {
 	}
 
-	public SaleDTO(Long id, Integer visited, Integer deals, Integer amount, LocalDate date, SellerDTO seller) {
+	public SaleDTO(Long id, Integer visited, Integer deals, Double amount, LocalDate date, SellerDTO seller) {
 		this.id = id;
 		this.visited = visited;
 		this.deals = deals;
@@ -59,11 +59,11 @@ public class SaleDTO {
 		this.deals = deals;
 	}
 
-	public Integer getAmount() {
+	public Double getAmount() {
 		return amount;
 	}
 
-	public void setAmount(Integer amount) {
+	public void setAmount(Double amount) {
 		this.amount = amount;
 	}
 
@@ -81,5 +81,5 @@ public class SaleDTO {
 
 	public void setSeller(SellerDTO seller) {
 		this.seller = seller;
-	}	
+	}
 }
