+++ backend/src/main/java/com/devsuperior/dsvendas/entities/Sale.java
@@ -13,13 +13,13 @@ import javax.persistence.Table;
 @Entity
 @Table(name = "tb_sales")
 public class Sale {
-	
+
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 	private Integer visited;
 	private Integer deals;
-	private Integer amount;
+	private Double amount;
 	private LocalDate date;
 	
 	@ManyToOne
@@ -29,7 +29,7 @@ public class Sale {
 	public Sale() {
 	}
 
-	public Sale(Long id, Integer visited, Integer deals, Integer amount, LocalDate date, Seller seller) {
+	public Sale(Long id, Integer visited, Integer deals, Double amount, LocalDate date, Seller seller) {
 		this.id = id;
 		this.visited = visited;
 		this.deals = deals;
@@ -62,11 +62,11 @@ public class Sale {
 		this.deals = deals;
 	}
 
-	public Integer getAmount() {
+	public Double getAmount() {
 		return amount;
 	}
 
-	public void setAmount(Integer amount) {
+	public void setAmount(Double amount) {
 		this.amount = amount;
 	}
